% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austraits.portal-package.R
\docType{package}
\name{austraits.portal-package}
\alias{austraits.portal}
\alias{austraits.portal-package}
\title{austraits.portal: Shiny interface to the AusTraits database}
\description{
Shiny R package to the AusTraits Plant Traits database.
}
\author{
\strong{Maintainer}: Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID})

Authors:
\itemize{
  \item Elizabeth Wenk \email{e.wenk@unsw.edu.au} (\href{https://orcid.org/0000-0001-5640-5910}{ORCID})
  \item Fonti Kar \email{f.kar@unsw.edu.au} (\href{https://orcid.org/0000-0002-2760-3974}{ORCID})
}

}
\keyword{internal}
