% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_means.R
\name{estimate_species_trait_means}
\alias{estimate_species_trait_means}
\title{Estimate Species Trait Means}
\usage{
estimate_species_trait_means(austraits)
}
\arguments{
\item{austraits}{A data frame or list containing trait data, typically in the Austraits format.}
}
\value{
A data frame with estimated mean trait values for each species.
}
\description{
Calculates mean trait values for each species in the provided Austraits dataset.
}
\details{
This function processes the input Austraits dataset to compute mean values of traits for each species.
}
\examples{
\dontrun{
means <- estimate_species_trait_means(austraits)
}

}
