% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_usage_text.R
\name{export_bibtex_for_data}
\alias{export_bibtex_for_data}
\title{Export BibTeX Entries for Data}
\usage{
export_bibtex_for_data(
  keys,
  filename,
  refs = RefManageR::ReadBib(file = "inst/extdata/austraits/sources.bib", check = FALSE,
    .Encoding = "UTF-8")
)
}
\arguments{
\item{keys}{A character vector of BibTeX entry keys to export.}

\item{filename}{A string specifying the path to the output file where the BibTeX entries will be saved.}

\item{refs}{A BibTeX object containing the references. This is typically read from a file.}
}
\value{
Invisibly returns \code{NULL}. The function is called for its side effect of writing to a file.
}
\description{
Exports BibTeX entries corresponding to the provided keys to a specified file.
}
\examples{
\dontrun{
export_bibtex_for_data(c("key1", "key2"), "output.bib")
}

}
