% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a-helper.R
\name{extract_distinct_values}
\alias{extract_distinct_values}
\title{Find distinct values for a given variable}
\usage{
extract_distinct_values(data, var_name)
}
\description{
Find distinct values for a given variable
}
\keyword{internal}
