% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trait_profile.R
\name{generate_trait_profile}
\alias{generate_trait_profile}
\title{Generate Text Summary for a trait}
\usage{
generate_trait_profile(data_trait)
}
\arguments{
\item{data_trait}{A data frame}

\item{trait}{A character string specifying the name of the trait to summarize.}
}
\value{
A character string containing the generated summary for the specified trait.
}
\description{
Creates a textual summary for a specified trait based on the provided data.
}
\examples{
\dontrun{
data_trait <- arrow::open_dataset("inst/extdata/austraits/austraits-6.0.0-mid-flatten.parquet") |> filter(trait_name == "leaf_area") |> collect()
generate_trait_profile(data_trait, "leaf_area")
}
}
