% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_usage_text.R
\name{generate_usage_and_citations_text}
\alias{generate_usage_and_citations_text}
\title{Generate Usage and Citations Text}
\usage{
generate_usage_and_citations_text(data)
}
\arguments{
\item{data}{A data object containing from which references are sourced, needed to generate usage and citation text.}
}
\value{
A character string with formatted usage and citation details.
}
\description{
Generates formatted text for usage instructions and citation information based on the selected data
}
\examples{
\dontrun{
  usage_text <- generate_usage_and_citations_text(my_data)
  cat(usage_text)
}
}
