% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_values.R
\name{plot_trait_distribution}
\alias{plot_trait_distribution}
\title{Plot Distribution of a Trait}
\usage{
plot_trait_distribution(data, trait, ...)
}
\arguments{
\item{data}{A data frame containing trait data.}

\item{trait}{A character string specifying the name of the categorical trait to plot.}

\item{...}{Additional arguments passed to the function.}
}
\description{
Generates a plot showing the distribution of a specified categorical trait across families.
}
